+++ app/detail/[id]/page.jsx
